requirements.txt
python-dotenv
pydantic
pydantic-settings
uvicorn
fastapi
